Add files via upload
Code cleaned. Cumulative plot issue fixed.
Note: Forward implies Reverse and vice versa. All processes are done as per this convention, and the labels interchanged at last before plotting to be consistent with reality. This was done since at the time it seemed easier than renaming the data which was interchanged while given to me.